fixed up routes
