--- controllers/mainRoutes.js
@@ -1,7 +1,0 @@
-const router = require('express').Router();
-
-router.get('/', (req, res) => {
- res.render('main'); 
-});
-
-module.exports = router;
